requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
motor>=3.7.0
pymongo[srv]>=4.10.0
python-dotenv>=1.1.0
